refactor: align tests with new `ValueWithUnit` hierarchy
